custodix-ai/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
oracledb
sqlalchemy